+++ pkg/pub_dartdoc/lib/src/pub_hooks.dart
@@ -17,8 +17,17 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:watcher/watcher.dart';
 
-const _maxFileCount = 10000;
-const _maxTotalLengthBytes = 10 * 1024 * 1024;
+const _defaultMaxFileCount = 10 * 1000 * 1000; // 10 million files
+const _defaultMaxTotalLengthBytes = 2 * 1024 * 1024 * 1024; // 2 GiB
+
+/// Thrown when current output exceeds limits.
+class DocumentationTooBigException implements Exception {
+  final String _message;
+  DocumentationTooBigException(this._message);
+
+  @override
+  String toString() => _message;
+}
 
 /// Creates an overlay file system with binary file support on top
 /// of the input sources.
@@ -28,12 +37,20 @@ const _maxTotalLengthBytes = 10 * 1024 * 1024;
 class PubResourceProvider implements ResourceProvider {
   final ResourceProvider _defaultProvider;
   final _memoryResourceProvider = MemoryResourceProvider();
+  final int _maxFileCount;
+  final int _maxTotalLengthBytes;
   bool _isSdkDocs = false;
   String _outputPath;
   int _fileCount = 0;
   int _totalLengthBytes = 0;
 
-  PubResourceProvider(this._defaultProvider);
+  PubResourceProvider(
+    this._defaultProvider, {
+    int maxFileCount,
+    int maxTotalLengthBytes,
+  })  : _maxFileCount = maxFileCount ?? _defaultMaxFileCount,
+        _maxTotalLengthBytes =
+            maxTotalLengthBytes ?? _defaultMaxTotalLengthBytes;
 
   /// Writes in-memory files to disk.
   void writeFilesToDiskSync() {
@@ -61,11 +78,11 @@ class PubResourceProvider implements ResourceProvider {
       return;
     }
     if (_fileCount > _maxFileCount) {
-      throw AssertionError(
+      throw DocumentationTooBigException(
           'Reached $_maxFileCount files in the output directory.');
     }
     if (_totalLengthBytes > _maxTotalLengthBytes) {
-      throw AssertionError(
+      throw DocumentationTooBigException(
           'Reached $_maxTotalLengthBytes bytes in the output directory.');
     }
   }