+++ pkg/pub_dartdoc/test/file_limit_test.dart
@@ -9,22 +9,22 @@ import 'package:pub_dartdoc/src/pub_hooks.dart';
 
 void main() {
   test('limit number of files', () {
-    final provider = PubResourceProvider(MemoryResourceProvider());
+    final provider =
+        PubResourceProvider(MemoryResourceProvider(), maxFileCount: 1000);
 
-    for (var i = 0; i < 10000; i++) {
+    for (var i = 0; i < 1000; i++) {
       provider.getFile('/tmp/$i.txt').writeAsStringSync('x');
     }
 
     expect(() => provider.getFile('/tmp/next.txt').writeAsStringSync('next'),
-        throwsA(isA<AssertionError>()));
+        throwsA(isA<DocumentationTooBigException>()));
   });
 
   test('limit total bytes', () {
-    final provider = PubResourceProvider(MemoryResourceProvider());
-    provider
-        .getFile('/tmp/1')
-        .writeAsBytesSync(List<int>.filled(10 * 1024 * 1024, 0));
+    final provider = PubResourceProvider(MemoryResourceProvider(),
+        maxTotalLengthBytes: 1024);
+    provider.getFile('/tmp/1').writeAsBytesSync(List<int>.filled(1024, 0));
     expect(() => provider.getFile('/tmp/2').writeAsBytesSync(<int>[0]),
-        throwsA(isA<AssertionError>()));
+        throwsA(isA<DocumentationTooBigException>()));
   });
 }